services/api_gateway/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "api-gateway-service"
version = "0.1.0"
description = "API Gateway Service for D&D Character Creator"
packages = [{include = "api_gateway", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
prometheus-client = "^0.17.0"
python-dotenv = "^1.0.0"
pydantic = {extras = ["email"], version = "^2.3.0"}
httpx = "^0.24.1"
structlog = "^23.1.0"
traefik = "^2.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.286"
mypy = "^1.5.1"
